# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SATBDS 2027 Symposium API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sponsors:
    get:
      operationId: getSponsors
      tags: [sponsors]
      summary: List all sponsors
      responses:
        "200":
          description: List of sponsors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sponsor"
    post:
      operationId: createSponsor
      tags: [sponsors]
      summary: Create a sponsor (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SponsorInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sponsor"
  /sponsors/{id}:
    put:
      operationId: updateSponsor
      tags: [sponsors]
      summary: Update a sponsor (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SponsorInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sponsor"
    delete:
      operationId: deleteSponsor
      tags: [sponsors]
      summary: Delete a sponsor (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Sponsor:
      type: object
      required: [id, name, tier]
      properties:
        id:
          type: integer
        name:
          type: string
        tier:
          type: string
          enum: [platinum, gold, silver, bronze]
        logoUrl:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
    SponsorInput:
      type: object
      required: [name, tier]
      properties:
        name:
          type: string
        tier:
          type: string
          enum: [platinum, gold, silver, bronze]
        logoUrl:
          type: string
        website:
          type: string
...