# lib/api-zod/src/generated/types/sponsor.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SATBDS 2027 Symposium API
 * OpenAPI spec version: 0.1.0
 */
import type { SponsorTier } from './sponsorTier';

export interface Sponsor {
  id: number;
  name: string;
  tier: SponsorTier;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  website?: string | null;
}